updated install script to support local installation

+++ install.sh
@@ -3,10 +3,18 @@
 if [ ! -f ~/Library/SyncDockerTime/io.github.sync-docker-time.plist ];
 then
   mkdir -p ~/Library/SyncDockerTime/log
-  curl -L -s https://github.com/x11vnc/sync-docker-time/archive/master.zip | \
-      bsdtar xf - -C ~/Library/SyncDockerTime/ --strip-components 1
-  rm -rf ~/Library/SyncDockerTime/sleepwatcher_2.2
-  chmod a+x ~/Library/SyncDockerTime/bin/*
+
+  if [[ "$0" -eq "bash" ]]; then
+    # Running in a pipe. Always download latest version
+    curl -L -s https://github.com/x11vnc/sync-docker-time/archive/master.zip | \
+        bsdtar xf - -C ~/Library/SyncDockerTime/ --strip-components 1
+    rm -rf ~/Library/SyncDockerTime/sleepwatcher_2.2
+    chmod a+x ~/Library/SyncDockerTime/bin/*
+  else
+    # Running locally
+    cp -r bin io.github.sync-docker-time.plist LICENSE uninstall.sh ~/Library/SyncDockerTime/
+  fi
+
   ln -s -f ~/Library/SyncDockerTime/io.github.sync-docker-time.plist \
       ~/Library/LaunchAgents/
   perl -e "s/\~/\/Users\/$USER/g" -p -i ~/Library/SyncDockerTime/io.github.sync-docker-time.plist
@@ -19,8 +27,8 @@ then
     launchctl list io.github.sync-docker-time | grep -q '"LastExitStatus" = 0'
 
   if [ $? -eq '0' ]; then
-    echo "sync-docker-time was installed successfully"
+    echo "SyncDockerTime was installed successfully"
   else
-    echo "Installation of sync-docker-time has failed"
+    echo "Installation of SyncDockerTime has failed"
   fi
 fi
